+++ include/Material.h
@@ -9,13 +9,13 @@ class Material
 {
 public:
     Color color = {Color::WHITE};
-    float ambient {0.1f};
-    float diffuse {0.9f};
-    float specular {0.9f};
-    float shininess {200.0f};
-    float reflective {0.0f};
-    float transparency {0.f};
-    float refraction_index {1.f};
+    double ambient {0.1f};
+    double diffuse {0.9f};
+    double specular {0.9f};
+    double shininess {200.0f};
+    double reflective {0.0f};
+    double transparency {0.f};
+    double refraction_index {1.f};
 
     std::shared_ptr<Pattern> pattern;
