+++ include/Transformation.h
@@ -5,11 +5,11 @@
 class Tuple;
 
 namespace matrix {
-Matrix4 Translation(float x, float y, float z);
-Matrix4 Scale(float x, float y, float z);
-Matrix4 RotationX(float rad);
-Matrix4 RotationY(float rad);
-Matrix4 RotationZ(float rad);
-Matrix4 Shear(float xy, float xz, float yx, float yz, float zx, float zy);
+Matrix4 Translation(double x, double y, double z);
+Matrix4 Scale(double x, double y, double z);
+Matrix4 RotationX(double rad);
+Matrix4 RotationY(double rad);
+Matrix4 RotationZ(double rad);
+Matrix4 Shear(double xy, double xz, double yx, double yz, double zx, double zy);
 Matrix4 ViewTransformation(Tuple from, Tuple to, Tuple up);
 }
